+++ src/LabelMePanel.ts
@@ -545,6 +545,12 @@ export class LabelMePanel {
                             <div id="canvasWrapper" class="canvas-wrapper">
                                 <canvas id="canvas"></canvas>
                                 <svg id="svgOverlay" class="svg-overlay"></svg>
+                                <div id="shapeContextMenu" class="shape-context-menu" style="display: none;">
+                                    <div class="context-menu-item" id="contextMenuEdit">Edit</div>
+                                    <div class="context-menu-item" id="contextMenuRename">Rename</div>
+                                    <div class="context-menu-item" id="contextMenuToggleVisible">Hide</div>
+                                    <div class="context-menu-item context-menu-danger" id="contextMenuDelete">Delete</div>
+                                </div>
                             </div>
                         </div>
                     </div>